\name{jointHWGraph-package}
\alias{jointHWGraph-package}
\alias{jointHWGraph}
\docType{package}
\title{
\packageTitle{jointHWGraph}
}
\description{
\packageDescription{jointHWGraph}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{jointHWGraph}

Maintainer: \packageMaintainer{jointHWGraph}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = jointHWGraph)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{jointHWGraph}}
%\section{Documentation Index}{\packageIndices{jointHWGraph}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
